tests/wpt/mozilla/tests/accessibility-tree/accessibility-update-partial-subtree-move-and-delete.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<section>
  <div id="div1">
    <div id="div2">
      <h1 id="h1">This is an h1</h1>
      <p id="p">This is a paragraph</p>
    </div>
  </div>
  <div id="div3"><h2 id="h2">This is an h2</h2></div>
</section>

<script>
  setup({ explicit_done: true });
  window.onload = () => {
    test((t) => {
      ServoTestUtils.ensureAccessibilityActive();
      let initial = ServoTestUtils.forceAccessibilityUpdate();

      assert_equals(initial.nodesUpdatedFromDom, 29, "nodesUpdatedFromDom");
      assert_equals(initial.nodesUpdatedFromTree, 29, "nodesUpdatedFromTree");
      assert_equals(initial.nodesInTreeUpdate, 29, "nodesInTreeUpdate");

      div3.moveBefore(div2, h2);
      div3.moveBefore(h1, div2);
      p.remove();
      div1.remove();

      let update = ServoTestUtils.forceAccessibilityUpdate();

      // Updated nodes:
      // - div1 (div2 removed - redundant update, since div1 is removed!)
      // - <section> (div1 removed)
      // - div2 (<p> and <h1> removed)
      // - div3 (new children)
      assert_equals(update.nodesUpdatedFromDom, 4, "nodesUpdatedFromDom");

      // <html>, <body>, <section>, div3, div2
      assert_equals(update.nodesUpdatedFromTree, 5, "nodesUpdatedFromTree");

      // Changed nodes:
      // - <section>, children changed
      // - div2, children changed
      // - div3, children changed
      assert_equals(update.nodesInTreeUpdate, 3, "nodesInTreeUpdate");
    }, "");

    done();
  };
</script>
